backend/pyproject.toml
[project]
name = "olympiad-preparation"
version = "7.1.0"
description = "Olympiad preparation app backend"
authors = [{ name = "mrdsx" }]
readme = "../README.md"
license = "../LICENSE.txt"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.121.0,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "google-genai (>=1.49.0,<2.0.0)",
    "async-lru (>=2.0.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
]


[tool.poetry.group.test.dependencies]
pytest = "^9.0.1"
pytest-asyncio = "^1.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
pythonpath = "src"
